perf(storage): bound the unbounded scans + caches (F-1346 / G11 cluster)
getCoinBySlug/getNativeCoin xlm_usd CTE gets the 24h floor the list
query documents as required (~40s scans + days-stale-as-current risk);
FindSorobanEventsLedgerGaps gets the ledger_close_time prune its sibling
uses (was full-hypertable scan); fx-resolver cache gets bounded eviction
+ a freshness-floored query; NUMERIC volume sums stay NUMERIC (/1e7
float coercion); InsertTrade registry-hook failures now logged not
silently swallowed.

+++ internal/storage/timescale/markets.go
@@ -337,11 +337,11 @@ func buildPoolsQuery(since time.Time, filter PoolsFilter, cursor string, limit i
               CASE
                 WHEN p.base_asset IN ('native', 'CAS3J7GYLGXMF6TDJBBYYSE3HQ6BBSMLNUQ34T6TZMYMW2EVH34XOWMA')
                   THEN COALESCE(SUM(p.sum_base_unpriced)
-                                FILTER (WHERE p.bucket >= NOW() - INTERVAL '24 hours'), 0) / 1e7
+                                FILTER (WHERE p.bucket >= NOW() - INTERVAL '24 hours'), 0) / 1e7::numeric
                        * COALESCE((SELECT vwap FROM xlm_usd), 0)
                 WHEN p.quote_asset IN ('native', 'CAS3J7GYLGXMF6TDJBBYYSE3HQ6BBSMLNUQ34T6TZMYMW2EVH34XOWMA')
                   THEN COALESCE(SUM(p.sum_quote_unpriced)
-                                FILTER (WHERE p.bucket >= NOW() - INTERVAL '24 hours'), 0) / 1e7
+                                FILTER (WHERE p.bucket >= NOW() - INTERVAL '24 hours'), 0) / 1e7::numeric
                        * COALESCE((SELECT vwap FROM xlm_usd), 0)
                 ELSE 0
               END
